Updated functions/s2_scl_weights.js (added 2 lines, deleted 2 lines).

+++ functions/s2_scl_weights.js
@@ -6,11 +6,11 @@ exports.s2_scl_weights = function(img_coll, AOI, scale_to_use){
     
     var set_scl_each_field = function(region){
       
-      var scl_value = image.reduceRegion({
+      var scl_value = image.toByte().reduceRegion({
         reducer: ee.Reducer.mode(),
         geometry: region.geometry(),
         scale: scale_to_use
-      }).getNumber("SCL").toInt();
+      }).getNumber("SCL").toByte();
       
       return ee.Image(scl_value).clipToBoundsAndScale({
         geometry: region.geometry(),
